* Move save/load methods into different module * Use turn_number instance variable
