+++ ruby_oop/hangman/lib/hangman.rb
@@ -3,10 +3,12 @@
 require 'yaml'
 require './random_word'
 require './display_text'
+require './manage_game_state'
 
 # this is the Hangman game class
 class Hangman
   include DisplayableText
+  include ManageGameState
 
   attr_reader :guessed_letters, :secret_word, :hidden_word_lines, :turn_number
 
@@ -14,74 +16,26 @@ class Hangman
     @secret_word = RandomWord.new.rand_word.chomp
     @guessed_letters = []
     @prompt_guess_letter = ''
-    @turn_number = 0
+    @turn_number = 1
   end
 
   def play
     load_game_prompt
     load_game if @prompt_response == 'Y'
-    # @secret_word
     puts "#{create_hidden_lines} Word to guess is #{@secret_word.length} letters long! #{@secret_word}"
-    until game_over?(turn_number)
-      each_turn(turn_number)
+    until game_over?
+      each_turn
       @turn_number += 1
-    end
-  end
-
-  def load_game_prompt
-    puts DisplayableText::LOAD_GAME_CHOICE
-    @prompt_response = gets.chomp.upcase
-    case @prompt_response
-    when 'Y'
-      puts DisplayableText::LOADED_GAME_CONFIRMATION
-      load_game
-    end
-  end
-
-  def load_game_play
-    puts @hidden_word_lines
-    until game_over?(turn_number)
-      each_turn(turn_number)
-      @turn_number += 1
-    end
-  end
-
-  def load_game
-    game_state = YAML.load(File.open('saved_game.yml', 'r'), permitted_classes: [Hangman, RandomWord])
-    pp game_state
-    @guessed_letters = game_state.guessed_letters
-    @secret_word = game_state.secret_word
-    @hidden_word_lines = game_state.hidden_word_lines
-    @turn_number = game_state.turn_number
-    game_state.load_game_play
-  end
-
-  def save_game_file
-    serialized_object = File.new('saved_game.yml', 'w')
-    serialized_object.puts YAML.dump(self)
-    serialized_object.close
-  end
-
-  def save_the_game
-    if @turn_number != 11
-      puts DisplayableText::ASK_TO_SAVE_GAME
-      response = gets.chomp
-      case response
-      when 'Y', 'y', 'yes', 'YES'
-        puts DisplayableText::GAME_SAVED_CONFIRMATION
-        save_game_file
-        exit(0)
-      end
     end
   end
 
   private
 
-  def display_turns_left(turn_number)
-    if turn_number == 11
+  def display_turns_left
+    if @turn_number == 11
       puts lose_text(@secret_word)
     else
-      puts turns_left(turn_number)
+      puts turns_left(@turn_number)
     end
   end
 
@@ -96,11 +50,11 @@ class Hangman
     @hidden_word_lines
   end
 
-  def each_turn(turn_number)
+  def each_turn
     prompt_for_letter
     display_guessed_letters
     puts reveal_letter_if_it_exists
-    display_turns_left(turn_number)
+    display_turns_left
     save_the_game
   end
 
@@ -114,8 +68,8 @@ class Hangman
     @prompt_guess_letter = gets.chomp.downcase
   end
 
-  def game_over?(turn_number)
-    return true if turn_number == 12
+  def game_over?
+    return true if @turn_number == 12
 
     @game_over = false
     if win?
